+++ rust/src/lib.rs
@@ -1,28 +1,17 @@
-use std::cmp::Ordering;
 use std::os::raw::{c_char};
 use std::ffi::{CString, CStr, c_void};
-use std::sync::Arc;
 use uuid::Uuid;
 
 use stack_epic_wallet_api::{self, Owner};
-use stack_epic_wallet_config::{WalletConfig, EpicboxConfig};
-use stack_epic_wallet_impls::{DefaultLCProvider, DefaultWalletImpl, HTTPNodeClient};
-
-use stack_epic_util::file::get_first_line;
-use stack_epic_wallet_util::epic_util::ZeroingString;
-use stack_epic_util::Mutex;
-use stack_epic_wallet_libwallet::{address, scan, wallet_lock, NodeClient, WalletInst, WalletLCProvider, Error};
+use stack_epic_wallet_config::EpicboxConfig;
+use stack_epic_wallet_libwallet::Error;
 use stack_epic_wallet_controller::Error as EpicWalletControllerError;
 
-use stack_epic_wallet_util::epic_keychain::{Keychain, ExtKeychain};
-
-use stack_epic_util::secp::key::{SecretKey, PublicKey};
-use stack_epic_util::secp::{Secp256k1};
+use stack_epic_util::secp::key::SecretKey;
 
 use android_logger::FilterBuilder;
 pub mod config;
 use crate::config::Config;
-use crate::config::create_wallet_config;
 pub mod mnemonic;
 
 use crate::mnemonic::mnemonic;
@@ -32,16 +21,19 @@ use crate::mnemonic::_get_mnemonic;
 pub mod wallet;
 
 use crate::wallet::Wallet;
+use crate::wallet::create_wallet;
+use crate::wallet::recover_from_mnemonic;
 use crate::wallet::open_wallet;
+use crate::wallet::get_wallet_info;
 use crate::wallet::validate_address;
+use crate::wallet::wallet_scan_outputs;
 use crate::wallet::tx_strategies;
 use crate::wallet::tx_create;
 use crate::wallet::txs_get;
 use crate::wallet::tx_cancel;
 use crate::wallet::delete_wallet;
 use crate::wallet::tx_send_http;
-use crate::wallet::WalletInfoFormatted;
-use crate::wallet::nano_to_deci;
+use crate::wallet::get_chain_height;
 
 pub mod listener;
 
@@ -962,384 +954,6 @@ fn _get_tx_fees(
     let p = s.as_ptr(); // Get a pointer to the underlaying memory for s
     std::mem::forget(s); // Give up the responsibility of cleaning up/freeing s
     Ok(p)
-}
-
-pub fn create_wallet(config: &str, phrase: &str, password: &str, name: &str) -> Result<String, Error> {
-    let wallet_pass = ZeroingString::from(password);
-    let wallet_config = match Config::from_str(&config) {
-        Ok(config) => {
-            config
-        }, Err(e) => {
-            return  Err(Error::from(EpicWalletControllerError::GenericError(format!(
-                "Error getting wallet config: {}",
-                e.to_string()
-            ))));
-        }
-    };
-
-    let wallet = match get_wallet(&wallet_config) {
-        Ok(wllet) => {
-            wllet
-        }
-        Err(e) => {
-            return  Err(e);
-        }
-    };
-    let mut wallet_lock = wallet.lock();
-    let lc = match wallet_lock.lc_provider() {
-        Ok(wallet_lc) => {
-            wallet_lc
-        }
-        Err(e) => {
-            return  Err(e);
-        }
-    };
-    let rec_phrase = ZeroingString::from(phrase);
-    let result = match lc.create_wallet(
-        Some(name),
-        Some(rec_phrase),
-        32,
-        wallet_pass,
-        false,
-    ) {
-        Ok(_) => {
-            "".to_string()
-        },
-        Err(e) => {
-            e.to_string()
-        },
-    };
-    Ok(result)
-}
-
-pub fn get_wallet_secret_key_pair(
-    wallet: &Wallet, keychain_mask: Option<SecretKey>, index: u32
-) -> Result<(SecretKey, PublicKey), Error>{
-    let parent_key_id = {
-        wallet_lock!(wallet, w);
-        w.parent_key_id().clone()
-    };
-    wallet_lock!(wallet, w);
-
-    let k = match w.keychain(keychain_mask.as_ref()) {
-        Ok(keychain) => {
-            keychain
-        }
-        Err(err) => {
-            return  Err(err);
-        }
-    };
-    let s = Secp256k1::new();
-    let sec_key = match address::address_from_derivation_path(
-        &k, &parent_key_id, index
-    ) {
-        Ok(s_key) => {
-            s_key
-        }
-        Err(err) => {
-            return Err(err);
-        }
-    };
-    let pub_key = match PublicKey::from_secret_key(&s, &sec_key) {
-        Ok(p_key) => {
-            p_key
-        }
-        Err(err) => {
-            return Err(Error::from(
-                EpicWalletControllerError::GenericError(
-                    format!("{}", err.to_string())
-                )
-            ));
-        }
-    };
-
-    Ok((sec_key, pub_key))
-}
-
-
-
-/*
-    Get wallet info
-    This contains wallet balances
-*/
-pub fn get_wallet_info(
-    wallet: &Wallet,
-    keychain_mask: Option<SecretKey>,
-    refresh_from_node: bool,
-    min_confirmations: u64
-) -> Result<WalletInfoFormatted, Error> {
-    let api = Owner::new(wallet.clone(), None);
-
-    match api.retrieve_summary_info(keychain_mask.as_ref(), refresh_from_node, min_confirmations) {
-        Ok((_, wallet_summary)) => {
-            Ok(WalletInfoFormatted {
-                last_confirmed_height: wallet_summary.last_confirmed_height,
-                minimum_confirmations: wallet_summary.minimum_confirmations,
-                total: nano_to_deci(wallet_summary.total),
-                amount_awaiting_finalization: nano_to_deci(wallet_summary.amount_awaiting_finalization),
-                amount_awaiting_confirmation: nano_to_deci(wallet_summary.amount_awaiting_confirmation),
-                amount_immature: nano_to_deci(wallet_summary.amount_immature),
-                amount_currently_spendable: nano_to_deci(wallet_summary.amount_currently_spendable),
-                amount_locked: nano_to_deci(wallet_summary.amount_locked)
-            })
-        }, Err(e) => {
-            return  Err(e);
-        }
-    }
-
-}
-
-/*
-    Recover wallet from mnemonic
-*/
-pub fn recover_from_mnemonic(mnemonic: &str, password: &str, config: &Config, name: &str) -> Result<(), Error> {
-    let wallet = match get_wallet(&config) {
-        Ok(conf) => {
-            conf
-        }
-        Err(e) => {
-            return  Err(e);
-        }
-    };
-    let mut w_lock = wallet.lock();
-    let lc = match w_lock.lc_provider() {
-        Ok(wallet_lc) => {
-            wallet_lc
-        }
-        Err(e) => {
-            return  Err(e);
-        }
-    };
-
-    //First check if wallet seed directory exists, if not create
-    if let Ok(exists_wallet_seed) = lc.wallet_exists(None) {
-        return if exists_wallet_seed {
-            match lc.recover_from_mnemonic(
-                ZeroingString::from(mnemonic), ZeroingString::from(password)
-            ) {
-                Ok(_) => {
-                    Ok(())
-                }
-                Err(e) => {
-                    Err(e)
-                }
-            }
-        } else {
-            match lc.create_wallet(
-                Some(&name),
-                Some(ZeroingString::from(mnemonic)),
-                32,
-                ZeroingString::from(password),
-                false,
-            ) {
-                Ok(_) => {
-                    Ok(())
-                }
-                Err(e) => {
-                    Err(e)
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-/*
-    Get wallet that will be used for calls to epic wallet
-*/
-fn get_wallet(config: &Config) -> Result<Wallet, Error> {
-    let wallet_config = match create_wallet_config(config.clone()) {
-        Ok(conf) => {
-            conf
-        } Err(e) => {
-            return Err(e);
-        }
-    };
-    let node_api_secret = get_first_line(wallet_config.node_api_secret_path.clone());
-    let node_client = HTTPNodeClient::new(&wallet_config.check_node_api_http_addr, node_api_secret).unwrap();
-    let wallet =  match inst_wallet::<
-        DefaultLCProvider<HTTPNodeClient, ExtKeychain>,
-        HTTPNodeClient,
-        ExtKeychain,
-    >(wallet_config.clone(), node_client) {
-        Ok(wallet_inst) => {
-            wallet_inst
-        }
-        Err(e) => {
-            return  Err(e);
-        }
-    };
-    return Ok(wallet);
-}
-/*
-    New wallet instance
-*/
-fn inst_wallet<L, C, K>(
-    config: WalletConfig,
-    node_client: C,
-) -> Result<Arc<Mutex<Box<dyn WalletInst<'static, L, C, K>>>>, Error>
-    where
-        DefaultWalletImpl<'static, C>: WalletInst<'static, L, C, K>,
-        L: WalletLCProvider<'static, C, K>,
-        C: NodeClient + 'static,
-        K: Keychain + 'static,
-{
-    let mut wallet = Box::new(DefaultWalletImpl::<'static, C>::new(node_client.clone()).unwrap())
-        as Box<dyn WalletInst<'static, L, C, K>>;
-    let lc = match wallet.lc_provider() {
-        Ok(wallet_lc) => {
-            wallet_lc
-        }
-        Err(err) => {
-            return  Err(err);
-        }
-    };
-    match lc.set_top_level_directory(&config.data_file_dir) {
-        Ok(_) => {
-            ()
-        }
-        Err(err) => {
-            return  Err(err);
-        }
-    };
-    Ok(Arc::new(Mutex::new(wallet)))
-}
-
-pub fn get_chain_height(config: &str) -> Result<u64, Error> {
-    let config = match Config::from_str(&config.to_string()) {
-        Ok(config) => {
-            config
-        }, Err(_e) => {
-            return Err(Error::from(EpicWalletControllerError::GenericError(format!(
-                "{}",
-                "Unable to get wallet config"
-            ))))
-        }
-    };
-    let wallet_config = match create_wallet_config(config.clone()) {
-        Ok(wallet_conf) => {
-            wallet_conf
-        }
-        Err(e) => {
-            return  Err(e);
-        }
-    };
-    let node_api_secret = get_first_line(wallet_config.node_api_secret_path.clone());
-    let node_client = HTTPNodeClient::new(&wallet_config.check_node_api_http_addr, node_api_secret);
-    let chain_tip = match node_client?.chain_height() {
-        Ok(tip) => {
-            tip
-        }
-        Err(err) => {
-            return  Err(err);
-        }
-    };
-    Ok(chain_tip.0)
-}
-
-
-/*
-
-*/
-pub fn wallet_scan_outputs(
-    wallet: &Wallet,
-    keychain_mask: Option<SecretKey>,
-    start_height: Option<u64>,
-    number_of_blocks_to_scan: Option<u64>
-) -> Result<String, Error> {
-    let tip = {
-        wallet_lock!(wallet, w);
-        match w.w2n_client().get_chain_tip() {
-            Ok(chain_tip) => {
-                chain_tip.0
-            },
-            Err(_e) => {
-                0
-            }
-        }
-    };
-
-    if tip == 0 {
-        return Err(Error::from(EpicWalletControllerError::GenericError(format!(
-            "{}",
-            "Unable to scan, could not determine chain height"
-        ))));
-    }
-
-    let start_height: u64 = match start_height {
-        Some(h) => h,
-        None => 1,
-    };
-
-    let number_of_blocks_to_scan: u64 = match number_of_blocks_to_scan {
-        Some(h) => h,
-        None => 0,
-    };
-
-    let last_block = start_height.clone() + number_of_blocks_to_scan;
-    let end_height: u64 = match last_block.cmp(&tip) {
-        Ordering::Less => {
-            last_block
-        },
-        Ordering::Greater => {
-            tip
-        },
-        Ordering::Equal => {
-            last_block
-        }
-    };
-
-    match scan(
-        wallet.clone(),
-        keychain_mask.as_ref(),
-        false,
-        start_height,
-        end_height,
-        &None
-    ) {
-        Ok(info) => {
-
-
-            let parent_key_id = {
-                wallet_lock!(wallet, w);
-                w.parent_key_id().clone()
-            };
-
-            {
-                wallet_lock!(wallet, w);
-                let mut batch = match w.batch(keychain_mask.as_ref()) {
-                    Ok(wallet_output_batch) => {
-                        wallet_output_batch
-                    }
-                    Err(err) => {
-                        return Err(err);
-                    }
-                };
-                match batch.save_last_confirmed_height(&parent_key_id, info.clone().height) {
-                    Ok(_) => {
-                        ()
-                    }
-                    Err(err) => {
-                        return  Err(err);
-                    }
-                };
-                match batch.commit() {
-                    Ok(_) => {
-                        ()
-                    }
-                    Err(err) => {
-                        return  Err(err);
-                    }
-                }
-            };
-
-
-            let result = info.height;
-            Ok(serde_json::to_string(&result).unwrap())
-        }, Err(e) => {
-            return  Err(e);
-        }
-    }
 }
 
 #[no_mangle]
